Fix a compile error on swift 5.4 or earlier
The error detail is here:

> Extensions.swift:11:18: error: cannot convert value of type 'CGFloat' to
> expected argument type 'Double'
>
>         let v = (self * p).rounded(.towardZero) / p
>                  ^
>                  Double( )

+++ Sources/Extensions.swift
@@ -7,7 +7,7 @@ import UIKit
 extension CGFloat {
     /// Returns this value rounded to an logical pixel value by a display scale
     func rounded(by displayScale: CGFloat) -> CGFloat {
-        let p = 1.0e9
+        let p = CGFloat(1.0e9)
         let v = (self * p).rounded(.towardZero) / p
         return (v * displayScale).rounded(.toNearestOrAwayFromZero) / displayScale
     }
